add a net_send_until_acked sending routine
in this patch, the net_poll function is rewritten not to directly call
net_proc, but to have a return value. thus, its main function (check if
an incoming package is "for us" and copy it to the receive buffer) can
also be used in net_send_until_acked, which drops everything except the
ack it is waiting for.